+++ raspberryServer/Arduino.h
@@ -71,7 +71,7 @@ class Arduino{
 		void setOccupancy(bool value);
 		double getRef();
 		double getPower(); // Nâo implementada
-		long getTime();
+		double getTime();
 		void send(string str);
 		void ledON(int pwm = 255);
 		void ledON(double V);
